: Опрос 7кт через модем
Task-Url:

+++ bundles/ru.7kt.core/src/ru/sevenkt/annotations/Address.java
@@ -8,5 +8,5 @@ import java.lang.annotation.Target;
 @Target(value=ElementType.FIELD)
 @Retention(value= RetentionPolicy.RUNTIME)
 public @interface Address {
-	long value();
+	int value();
 }
